Se modifico layout de detalle de empresa
