--- app/src/main/java/Entidades/Empresa.kt
@@ -1,12 +1,0 @@
-package Entidades
-
-data class Empresa (var titulo_empresa:String, var tipoempresa: String)
-{
-    init {
-         this.titulo_empresa = titulo_empresa
-        this.tipoempresa = tipoempresa
-    }
-
-
-
-}
