+++ app/src/main/java/com/uabc/edu/gnomon/ui/favoritos/favoritos.kt
@@ -1,34 +1,21 @@
 package com.uabc.edu.gnomon.ui.favoritos
 
-import Entidades.Empresa
 import adaptadores.AdapterEmpresa
-import android.content.Context
-import android.content.Intent
 import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.webkit.WebChromeClient
-import android.webkit.WebViewClient
-import android.support.v4.app.*
-import android.widget.LinearLayout
-import android.widget.RelativeLayout
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.uabc.edu.gnomon.MainActivity
 import com.uabc.edu.gnomon.R
-import com.uabc.edu.gnomon.ui.perfil.perfil
-import kotlinx.android.synthetic.*
-import kotlinx.android.synthetic.main.fragment_empresa.*
-import kotlinx.android.synthetic.main.lista_favoritos.*
+import com.uabc.edu.gnomon.ui.empresa.Detalle_Fragment
 
 class favoritos () : Fragment(),AdapterEmpresa.OnFavoritoListener {
 
-
     private lateinit var recyclerView: RecyclerView
-    private val BASE_URL = "http://test.beetsool.com/gnomon/index.html"
+
     companion object {
        fun newInstance() = favoritos()
 
@@ -57,19 +44,9 @@ class favoritos () : Fragment(),AdapterEmpresa.OnFavoritoListener {
         recyclerView.setHasFixedSize(true)
         val layoutManager = LinearLayoutManager(context)
         recyclerView.layoutManager = layoutManager
-        recyclerView.adapter = AdapterEmpresa(myDataSet)
+        recyclerView.adapter = AdapterEmpresa(myDataSet, this)
 
-        webView.webChromeClient = object : WebChromeClient() {
 
-        }
-
-        webView.webViewClient = object :WebViewClient(){
-
-        }
-
-        val settings = webView.settings
-        settings.javaScriptEnabled = true
-        webView.loadUrl(BASE_URL)
 
         return view
 
@@ -86,7 +63,7 @@ class favoritos () : Fragment(),AdapterEmpresa.OnFavoritoListener {
     }
 
     override fun onFavoritoClick(position: Int) {
-        val fragment = perfil()
+        val fragment = Detalle_Fragment()
         val fragmentManager = requireActivity().supportFragmentManager
         val fragmentTransaction = fragmentManager.beginTransaction()
         fragmentTransaction.replace(R.id.nav_host_fragment, fragment)
