+++ app/src/main/res/layout/fragment_detalle.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="620dp"
+        android:orientation="vertical">
+
+        <WebView
+            android:id="@+id/wvContainer"
+            android:layout_width="422dp"
+            android:layout_height="460dp"
+            android:layout_alignParentTop="true"
+            android:layout_weight="9.0"></WebView>
+
+
+        <LinearLayout
+            android:layout_width="fill_parent"
+            android:layout_height="65dp"
+            android:layout_alignParentBottom="true"
+            android:orientation="horizontal">
+
+            <Button
+                android:id="@+id/button2"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:background="@color/colorPrimaryDark"
+                android:text="AGENDAR"
+                android:textColor="#FFFDFD" />
+
+
+        </LinearLayout>
+
+    </LinearLayout>
+</RelativeLayout>
+
