+++ app/src/main/res/layout/fragment_item.xml
@@ -2,6 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
+    android:background="#E3E3E3"
     android:orientation="horizontal">
 
     <TextView
