+++ app/src/main/res/layout/fragment_notifications.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#FFFFFF"
     tools:context=".ui.notifications.NotificationsFragment">
 
     <TextView
