+++ app/src/main/res/layout/item_text_view.xml
@@ -6,6 +6,7 @@
     card_view:cardUseCompatPadding="true"
     card_view:cardElevation="4dp"
     card_view:cardCornerRadius="3dp"
+    android:background="#FFFFFF"
     android:layout_margin="6dp">
 
     <LinearLayout
