+++ app/src/main/res/layout/lista_favoritos.xml
@@ -8,6 +8,7 @@
     android:orientation="vertical"
     app:cardCornerRadius="3dp"
     app:cardElevation="3dp"
+    android:background="#FFFFFF"
     app:cardUseCompatPadding="true">
 
     <RelativeLayout
