+++ app/src/main/res/layout/mensajes_fragment.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#FFFFFF"
     tools:context=".ui.mensajes.mensajes">
 
     <TextView
