+++ app/src/main/res/values/strings.xml
@@ -25,4 +25,6 @@
     <string name="registrar">Registrar</string>
     <string name="dummy_button">Dummy Button</string>
     <string name="dummy_content">DUMMY\nCONTENT</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
